# schemas/ripd/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RIP Agent Server",
    "description": "Agent session control plane (HTTP/SSE).",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/{id}/cancel": {
      "post": {
        "operationId": "cancel_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session canceled"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/events": {
      "get": {
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of event frames"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/input": {
      "post": {
        "operationId": "send_input",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Input accepted"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "operationId": "list_tasks",
        "responses": {
          "200": {
            "description": "List tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskStatusResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSpawnPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid task request"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "task_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/cancel": {
      "post": {
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCancelPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Cancel requested"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/events": {
      "get": {
        "operationId": "stream_task_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of task event frames"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/output": {
      "get": {
        "operationId": "task_output",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetch task output (range)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOutputResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/resize": {
      "post": {
        "operationId": "task_resize",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskResizePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Resize accepted"
          },
          "400": {
            "description": "Invalid resize request"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/signal": {
      "post": {
        "operationId": "task_signal",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSignalPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Signal accepted"
          },
          "400": {
            "description": "Invalid signal request"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/tasks/{id}/stdin": {
      "post": {
        "operationId": "task_write_stdin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskWriteStdinPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Stdin accepted"
          },
          "400": {
            "description": "Invalid stdin request"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
...
    "/threads/{id}": {
      "get": {
        "operationId": "thread_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadMeta"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/branch": {
      "post": {
        "operationId": "thread_branch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Parent thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadBranchPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadBranchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid branch request"
          },
          "404": {
            "description": "Thread or branch point not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-auto": {
      "post": {
        "operationId": "thread_compaction_auto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionAutoV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Auto-compaction no-op / dry-run result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoV1Response"
                }
              }
            }
          },
          "202": {
            "description": "Auto-compaction job spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid auto-compaction request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-auto-schedule": {
      "post": {
        "operationId": "thread_compaction_auto_schedule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionAutoScheduleV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scheduler no-op / dry-run / skipped result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoScheduleV1Response"
                }
              }
            }
          },
          "202": {
            "description": "Scheduler started a job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionAutoScheduleV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid schedule request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-checkpoint": {
      "post": {
        "operationId": "thread_compaction_checkpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCompactionCheckpointPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Compaction checkpoint created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadCompactionCheckpointResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid checkpoint request"
          },
          "404": {
            "description": "Thread or cut point not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-cut-points": {
      "post": {
        "operationId": "thread_compaction_cut_points",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionCutPointsV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Computed cut points",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionCutPointsV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid cut point request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/compaction-status": {
      "post": {
        "operationId": "thread_compaction_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompactionStatusV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compaction status projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompactionStatusV1Response"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status request"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/context-selection-status": {
      "post": {
        "operationId": "thread_context_selection_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextSelectionStatusV1Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Context selection strategy decisions (truth-derived)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextSelectionStatusV1Response"
                }
              }
            }
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/events": {
      "get": {
        "operationId": "thread_stream_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of thread event frames"
          },
          "404": {
            "description": "Thread not found"
          }
        }
      }
    },
    "/threads/{id}/handoff": {
      "post": {
        "operationId": "thread_handoff",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Source thread id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadHandoffPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Handoff thread created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadHandoffResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid handoff request"
          },
          "404": {
            "description": "Thread or handoff point not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiToolTaskExecutionMode": {
        "type": "string",
        "enum": [
          "pipes",
          "pty"
        ]
      },
...
      "InputPayload": {
        "type": "object",
        "required": [
          "input"
        ],
        "properties": {
          "input": {
            "type": "string"
          }
        }
      },
...
      "TaskCreated": {
        "type": "object",
        "required": [
          "task_id"
        ],
        "properties": {
          "task_id": {
            "type": "string"
          }
        }
      },
...
      "TaskSpawnPayload": {
        "type": "object",
        "required": [
          "tool"
        ],
        "properties": {
          "args": {},
          "execution_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ApiToolTaskExecutionMode"
              }
            ]
          },
          "origin_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tool": {
            "type": "string"
          }
        }
      },
      "TaskStatusResponse": {
        "type": "object",
        "required": [
          "task_id",
          "status",
          "tool",
          "execution_mode"
        ],
        "properties": {
          "artifacts": {},
          "ended_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ApiToolTaskExecutionMode"
          },
          "exit_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "started_at_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "$ref": "#/components/schemas/ApiToolTaskStatus"
          },
          "task_id": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "tool": {
            "type": "string"
          }
        }
      },
...
    }
  }
}